backend/requirements.txt
pip
autopep8
google-genai
python-dotenv
Flask==3.0.3
flask-cors
gunicorn==22.0.0
Werkzeug==3.0.6